// packages/shared/chat-ui/src/types.ts
export type DisplayContent = string | any[]
// ...
export enum ContentPartType {
  Text  = 'text',
  Image = 'image',
  Audio = 'audio',
}

export interface DisplayPart {
  type: ContentPartType
  text?: string
  url?: string
}